+++ app/javascript/app/pages/ndc-country/ndc-country.js
@@ -75,8 +75,7 @@ class NDCCountryContainer extends PureComponent {
       ...this.props,
       onSearchChange: this.onSearchChange,
       handleDropDownChange: this.handleDropDownChange,
-      handleCountryLink: this.handleCountryLink,
-      goBack: this.props.history.goBack
+      handleCountryLink: this.handleCountryLink
     });
   }
 }
@@ -84,7 +83,7 @@ class NDCCountryContainer extends PureComponent {
 NDCCountryContainer.propTypes = {
   history: Proptypes.object.isRequired,
   location: Proptypes.object.isRequired,
-  country: Proptypes.object.isRequired
+  country: Proptypes.object
 };
 
 export default withRouter(connect(mapStateToProps, null)(NDCCountryContainer));
